+++ BuddySystem2/.idea/.name
@@ -1,0 +1,1 @@
+My Application
