+++ MyApplication2/app/proguard-rules.pro
