+++ MyApplication2/app/src/debug/res/values/google_maps_api.xml
@@ -1,0 +1,19 @@
+<resources>
+    <!--
+    TODO: Before you run your application, you need a Google Maps API key.
+
+    To get one, follow this link, follow the directions and press "Create" at the end:
+
+    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=59:BD:5C:F3:BA:F4:E1:62:5D:5F:99:E9:0E:DA:6E:CB:8C:A7:67:95%3Bcom.example.wardw.myapplication
+
+    You can also add your credentials to an existing key, using this line:
+    59:BD:5C:F3:BA:F4:E1:62:5D:5F:99:E9:0E:DA:6E:CB:8C:A7:67:95;com.example.wardw.myapplication
+
+    Alternatively, follow the directions here:
+    https://developers.google.com/maps/documentation/android/start#get-key
+
+    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
+    string in this file.
+    -->
+    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyDGMWUPPu6l35zJMUhrHssvr-GzAGRlq0w</string>
+</resources>
