+++ MyApplication2/app/src/main/AndroidManifest.xml
