+++ MyApplication2/app/src/main/res/values/strings.xml
