+++ MyApplication2/gradle.properties
