+++ MyApplication2/gradle/wrapper/gradle-wrapper.properties
